delete path pha
